[aiecc] Fix stack-analysis correctness bugs; move it to a linkable library
Correctness fixes found auditing the call-graph stack analysis and its
driver integration:

- A relocation against the null symbol index (R_*_NONE, or a purely
  section-relative one) has no symbol to dereference, but both relocation
  loops in addObjectToStackGraph did `*rel.getSymbol()` unconditionally --
  undefined behavior on any real object carrying one. Guarded at both
  sites.

- A build that fails the post-build stack_size sufficiency check left the
  xclbin/insts this run's engine.run() had already written looking
  complete. A caller that doesn't check aiecc's exit code (or `make`, on
  the next invocation, seeing an up-to-date target) would pick up a binary
  whose buffers were placed against a stack_size now proven insufficient.
  Engine::run gained an optional out-param collecting every path it wrote
  for the requested outputs; the post-build check now unlinks them on
  failure.

- An explicit-but-provably-insufficient stack_size only ever warned,
  while an absent one hard-failed for the identical shortfall -- the
  asymmetry was backwards, since the post-build check has the TRUE
  requirement (not the early check's lower bound), so a warning there
  would ship a proven overflow. Both cases are now a hard failure;
  renamed checkAbsentStackSizeIsSufficient to checkStackSizeIsSufficient
  to match.

- The post-build check re-derives placement from a fresh parse, which
  `--placer=sa_placer --sa-seed=0` (explicitly documented as
  non-deterministic) can silently fail to reproduce, causing the object
  path lookup to miss and the check to no-op with no diagnostic. Now
  warns and skips instead.

- That same re-derivation re-ran the early pipeline stages and re-emitted
  every diagnostic already shown once during the real build. Wrapped in a
  ScopedDiagnosticHandler that suppresses them; only the new
  insufficient-stack_size diagnostics are shown.

- resolveIndirectCallEdges accumulated callee lists from hash-ordered
  maps without deduping, so the cycle/unmeasurable-symbol diagnostic
  (built from the first failing DFS path) could report different wording
  run to run for the same input. Now sorts and dedupes each node's callee
  list.

- graph.nodes is keyed by name alone; two objects defining a same-named
  static function would let the later one silently overwrite the
  earlier's frame size, which can undercount -- the one direction this
  analysis must never be wrong in. Now keeps the max.

- Unguarded integer arithmetic (ULEB128 frame sizes, the recursive
  max-path sum) could wrap negative on a malformed or hostile object,
  silently undercounting. Both now detect overflow and report the object
  unmeasurable instead.

- measureObjectDataSectionBytes summed section sizes without alignment
  padding, under-counting a reservation that is supposed to be a safe
  upper bound. Now aligns the running total to each section's own
  alignment before adding its size.

- The core-frame symbol name ("core_<col>_<row>") was independently
  reconstructed in aiecc.cpp and defined in AIECoreToStandard.cpp; moved
  the single definition to AIECoreMemory.h so both agree by construction.
  The aiecc.computed_stack_requirement attribute name is likewise now one
  shared constant instead of three copies of the same string literal.

- checkStackSizeRequirements's per-core scan re-parsed a shared
  link_files object once per core purely to collect its defined-function
  names (pass 1); that result depends only on the object, so it is now
  cached across cores in the same call. Pass 2 (closure-dependent) is
  deliberately left uncached and the reason documented.

- The reserved_data_size and stack_size skipped-artifact warnings had
  copy-pasted comma-joining logic; factored into appendSkippedArtifacts.
  The three near-identical "run this pass manager or report which stage
  failed" blocks in the post-build check collapsed into one runStage
  helper.

New: `--default-stack-size` lets a design say "assume N bytes for any
core that leaves stack_size absent" instead of the target's built-in
default, via a populateDefaultStackSize pass wired into both the main
build and the post-build re-derivation (so the "assumed" value in
diagnostics stays consistent between them). Explicit stack_size is
untouched either way.

Moves the analysis itself from tools/aiecc/StackSizeAnalysis.h into
lib/Analysis/ + include/aie/Analysis/, wired into CMake as its own
AIEAnalysis library: it reads object files, not MLIR, so it's linkable
and independently testable rather than living only inside the aiecc
driver.

Also corrects two stale/contradictory pieces of documentation this
branch introduced: AIEOps.td's CoreOp description still said the stack
reservation "is not checked" and never mentioned reserved_data_size at
all; IRTransforms.h's comment said a cycle *or* an unmeasurable symbol
both fail the build, when only a cycle does. Error messages now name the
IRON spelling (Worker(stack_size=...)) alongside the MLIR one.

New regression coverage: stack_size_explicit_insufficient_error.mlir and
default_stack_size.mlir (the two behavior changes above),
stack_size_indirect_call.mlir (function-pointer-table inference, verified
via llvm-readelf that no direct relocation exists between caller and
callee), stack_size_cross_object.mlir (the two-pass cross-object
attribution the file comment describes but nothing exercised), and
mixed_measurable_link_files_warns.mlir (a core with one measurable and
one unmeasurable-but-unreached link_files object, for both
reserved_data_size and stack_size -- previously only the
all-unmeasurable case was tested for stack_size, and reserved_data_size
had no such test at all).
