+++ python/utils/compile/utils.py
@@ -330,12 +330,9 @@ def compile_cxx_core_function(
             f"--target={target_arch}-none-unknown-elf",
         ]
         if not inline:
-            # Record each function's frame size in a `.stack_sizes` section,
-            # matching the flag aiecc's own core-object `llc` invocation
-            # passes (tools/aiecc/aiecc.cpp). Without this, kernel objects
-            # carry no stack accounting at all -- and kernels are where large
-            # frames actually live. Only meaningful for object codegen; the
-            # inline path emits textual LLVM IR with no frame layout yet.
+            # Matches the -stack-size-section aiecc's own core-object llc
+            # invocation passes; without it, kernel objects (where large
+            # frames actually live) carry no stack accounting at all.
             cmd.append("-fstack-size-section")
 
     # Add include directories
